Combining The Files
Added the python code to combine the files. I ran it on the WRF winter data. I think I need to run it on the other WRF data as well. I will check and update this later!

One of the ARCGIS functions does not seem to work if the code is run outside of ARCGIS itself. I am not sure what the problem is running it just in python.(arcpy.management.XYTableToPoint)

The final file is not uploadable to github since it is over 100 Mb. I will push it to the VAIA server.